+++ src/main/java/com/example/crudAeropuerto/Repository/UsuarioRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepo extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepo extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByEmail(String email);
 }
