+++ src/main/java/com/example/crudAeropuerto/Service/UsuarioServiceIMPL/UsuarioServiceIMPL.java
@@ -31,11 +31,11 @@ public class UsuarioServiceIMPL implements UsuarioService {
 
     @Override
     public Usuario BuscaUsuario(int idUsuario) {
-        return this.repo.findById((long) idUsuario).orElse(null);
+        return this.repo.findById((Integer) idUsuario).orElse(null);
     }
 
     @Override
     public void EliminaUsuario(int idUsuario) {
-        this.repo.deleteById((long) idUsuario);
+        this.repo.deleteById((Integer) idUsuario);
     }
 }
